Jenkins/scenarios/03-eks-deployment/requirements.txt
pytest>=7.0.0
requests>=2.28.0
boto3>=1.26.0
PyYAML>=6.0
